赠送积分将支持"固定值"与"按支付额百分比"二种方式

+++ application/common/model/Goods.php
@@ -67,7 +67,7 @@ class Goods extends CareyShop
         'is_postage'        => 'integer',
         'market_price'      => 'float',
         'shop_price'        => 'float',
-        'give_integral'     => 'integer',
+        'give_integral'     => 'float',
         'is_integral'       => 'integer',
         'least_sum'         => 'integer',
         'purchase_sum'      => 'integer',
